Testing: exclude browser utils from coverage
