+++ setup.cfg
@@ -29,4 +29,4 @@ diff=True
 
 [coverage:run]
 branch = True
-omit = */migrations/*
+omit = */migrations/*,*/testing/browser_utils*
