+++ shuup/testing/utils.py
@@ -14,9 +14,6 @@ from django.utils.module_loading import import_string
 from django.utils.translation import activate, get_language
 
 from shuup.admin.shop_provider import set_shop
-from shuup.admin.utils.tour import set_tour_complete
-from shuup.core import cache
-from shuup.testing.factories import get_default_shop
 
 
 def apply_request_middleware(request, **attrs):
@@ -108,45 +105,3 @@ def apply_all_middleware(request, **attrs):
     for key, value in attrs.items():
         setattr(request, key, value)
     return request
-
-
-def initialize_front_browser_test(browser, live_server):
-    activate("en")
-    get_default_shop()
-    url = live_server + "/"
-    browser.visit(url)
-    # set shop language to eng
-    browser.find_by_id("language-changer").click()
-    browser.find_by_xpath('//a[@class="language"]').first.click()
-    return browser
-
-
-def initialize_admin_browser_test(browser, live_server, settings, username="admin", password="password",
-                                  onboarding=False, language="en", shop=None, tour_complete=True):
-    if not onboarding:
-        settings.SHUUP_SETUP_WIZARD_PANE_SPEC = []
-    activate("en")
-    cache.clear()
-
-    shop = shop or get_default_shop()
-
-    if tour_complete:
-        from django.contrib.auth import get_user_model
-        user = get_user_model().objects.get(username=username)
-        set_tour_complete(shop, "dashboard", True, user)
-        set_tour_complete(shop, "home", True, user)
-        set_tour_complete(shop, "product", True, user)
-        set_tour_complete(shop, "category", True, user)
-
-    url = live_server + "/sa"
-    browser.visit(url)
-    browser.fill('username', username)
-    browser.fill('password', password)
-    browser.find_by_css(".btn.btn-primary.btn-lg.btn-block").first.click()
-
-    if not onboarding:
-        # set shop language to eng
-        browser.find_by_id("dropdownMenu").click()
-        browser.find_by_xpath('//a[@data-value="%s"]' % language).first.click()
-
-    return browser
