+++ shuup_tests/browser/admin/test_dev_onboarding.py
@@ -11,12 +11,11 @@ import pytest
 
 from django.core.management import call_command
 
-from shuup import configuration
 from shuup.core.models import AnonymousContact, Product, Shop, Supplier
 from shuup.testing.browser_utils import (
     click_element, wait_until_appeared, wait_until_condition
 )
-from shuup.testing.utils import initialize_admin_browser_test
+from shuup.testing.browser_utils import initialize_admin_browser_test
 
 pytestmark = pytest.mark.skipif(os.environ.get("SHUUP_BROWSER_TESTS", "0") != "1", reason="No browser tests run.")
