+++ shuup_tests/browser/admin/test_menu.py
@@ -14,7 +14,7 @@ from django.core.urlresolvers import reverse
 
 from shuup.testing.browser_utils import wait_until_condition, wait_until_appeared
 from shuup.testing.factories import get_default_shop
-from shuup.testing.utils import initialize_admin_browser_test
+from shuup.testing.browser_utils import initialize_admin_browser_test
 
 pytestmark = pytest.mark.skipif(os.environ.get("SHUUP_BROWSER_TESTS", "0") != "1", reason="No browser tests run.")
