+++ shuup_tests/browser/admin/test_order_creator.py
@@ -23,7 +23,7 @@ from shuup.testing.factories import (
     get_default_shipping_method, get_default_shop, get_default_supplier,
     get_initial_order_status
 )
-from shuup.testing.utils import initialize_admin_browser_test
+from shuup.testing.browser_utils import initialize_admin_browser_test
 
 pytestmark = pytest.mark.skipif(os.environ.get("SHUUP_BROWSER_TESTS", "0") != "1", reason="No browser tests run.")
 OBJECT_CREATED_LOG_IDENTIFIER = "object_created_signal_handled"
