+++ shuup_tests/browser/admin/test_order_detail.py
@@ -17,7 +17,7 @@ from shuup.testing.browser_utils import (
     click_element, wait_until_appeared, wait_until_condition
 )
 from shuup.testing.factories import create_empty_order, get_default_shop
-from shuup.testing.utils import initialize_admin_browser_test
+from shuup.testing.browser_utils import initialize_admin_browser_test
 
 pytestmark = pytest.mark.skipif(os.environ.get("SHUUP_BROWSER_TESTS", "0") != "1", reason="No browser tests run.")
