+++ shuup_tests/browser/admin/test_picotable.py
@@ -19,7 +19,7 @@ from shuup.testing.factories import (
     create_product, create_random_person, get_default_shop,
     get_default_supplier
 )
-from shuup.testing.utils import initialize_admin_browser_test
+from shuup.testing.browser_utils import initialize_admin_browser_test
 
 pytestmark = pytest.mark.skipif(os.environ.get("SHUUP_BROWSER_TESTS", "0") != "1", reason="No browser tests run.")
