+++ shuup_tests/browser/admin/test_product_detail.py
@@ -14,7 +14,7 @@ from shuup.testing.browser_utils import (
     click_element, wait_until_condition, wait_until_appeared
 )
 from shuup.testing.factories import create_product, get_default_shop
-from shuup.testing.utils import initialize_admin_browser_test
+from shuup.testing.browser_utils import initialize_admin_browser_test
 
 pytestmark = pytest.mark.skipif(os.environ.get("SHUUP_BROWSER_TESTS", "0") != "1", reason="No browser tests run.")
