+++ shuup_tests/browser/admin/test_refunds.py
@@ -20,7 +20,7 @@ from shuup.testing.factories import (
     create_order_with_product, get_default_product, get_default_shop,
     get_default_supplier
 )
-from shuup.testing.utils import initialize_admin_browser_test
+from shuup.testing.browser_utils import initialize_admin_browser_test
 from shuup.utils.i18n import format_money
 
 pytestmark = pytest.mark.skipif(os.environ.get("SHUUP_BROWSER_TESTS", "0") != "1", reason="No browser tests run.")
