+++ shuup_tests/browser/front/test_checkout_with_login_and_register.py
@@ -23,7 +23,7 @@ from shuup.testing.factories import (
     create_product, get_default_payment_method, get_default_shipping_method,
     get_default_shop, get_default_supplier
 )
-from shuup.testing.utils import initialize_front_browser_test
+from shuup.testing.browser_utils import initialize_front_browser_test
 
 pytestmark = pytest.mark.skipif(os.environ.get("SHUUP_BROWSER_TESTS", "0") != "1", reason="No browser tests run.")
