+++ shuup_tests/browser/front/test_gdpr_consent.py
@@ -15,7 +15,7 @@ from shuup.testing.browser_utils import (
     click_element, wait_until_appeared, wait_until_condition
 )
 from shuup.testing.factories import get_default_shop
-from shuup.testing.utils import initialize_front_browser_test
+from shuup.testing.browser_utils import initialize_front_browser_test
 
 pytestmark = pytest.mark.skipif(os.environ.get("SHUUP_BROWSER_TESTS", "0") != "1", reason="No browser tests run.")
