+++ shuup_tests/browser/front/test_product_view.py
@@ -18,7 +18,7 @@ from shuup.testing.factories import (
     create_product, get_default_category, get_default_shop,
     get_default_supplier
 )
-from shuup.testing.utils import initialize_front_browser_test
+from shuup.testing.browser_utils import initialize_front_browser_test
 
 pytestmark = pytest.mark.skipif(os.environ.get("SHUUP_BROWSER_TESTS", "0") != "1", reason="No browser tests run.")
