+++ shuup_tests/browser/notify/test_script_template.py
@@ -28,7 +28,7 @@ from shuup.testing.browser_utils import (
     click_element, move_to_element, wait_until_condition
 )
 from shuup.testing.notify_script_templates import DummyScriptTemplate
-from shuup.testing.utils import initialize_admin_browser_test
+from shuup.testing.browser_utils import initialize_admin_browser_test
 
 pytestmark = pytest.mark.skipif(os.environ.get("SHUUP_BROWSER_TESTS", "0") != "1", reason="No browser tests run.")
